Move run_command into test.helpers
